Work to clean up and simplify the AtomPub API significantly.
URLs are now structured like this:

A workspace: /WORKSPACE (can be POSTed to now - no more X-Workspace header)
Each workspace is now an atompub collection in and of itself.

An artifact's data: /WORKASPACE/ARTIFACT

An Atom Entry representation of an artifact: /WORKSPACE/ARTIFACT;entry

Version history: /WORKSPACE/ARTIFACT;history

A previous version:  /WORKSPACE/ARTIFACT?version=VERSION

A list of workspaces inside a workspace: /WORKSPACE;workspaces
This is an AtomPub collection that can be POSTed/DELETEd to to add/delete workspaces.

Along with this, I've expanded the metadata included in the feed and
added a bunch more tests for feeds. Next up:
1. I don't really test the case where Galaxy is not the root webapp. This
is probably broken right now, but I'm hoping to fix it tomorrow. It should 
just be a couple tweaks.
2. Modifying the metadata and lifecycle via the AtomPub API.
3. OpenSearch support