GALAXY-107: delete user capabilities
Fix problem whereby you could have workspaces and artifacts with the same
name inside a workspace.
(Hopefully this fixes the concurrency issue as well)

Return a 409 if the workspace/artifact already exists.

Add a logging.properties so the web-standalone spits out more logging
(Need to still fully test this in a bit. I don't understand this logging
mess we've got ourselves into in Java so if this is wrong let me know).