Start implementing access checks inside Galaxy.
Move Permission to an enum.

Split out the acegi security configuration so we can do
access checks in the core.

Make Artifact and Workspace inherit from Item so its easy
to pass in any object to do an access check. This also lays
the groundwork for making a Service class which extends Item
which can be stored in workspaces.