Rework the history/structure of everything so we're no longer fighting the GWT way. As an added result we have simplified code and we can actually create hard links to pages now!
In summary, every page is now a page in and of itself. None of this trying to only
change the sub-page within the RegistryPanel/AdministrationPanel. Instead panels
(i.e. BrowsePanel, WorkspaceForm, etc) embed a MenuPanel inside them. This makes
it much easier to create customized links for each page.

When you wish to navigate pages now you should use a Hyperlink with a history
token that has the necessary information inside it for the page to instantiate itself.
Parameters from the hyperlink will be passed into the onShow(List) method.

I also reworked all the forms as well so they inherit from AbstractForm which simplify things.

I think the outstanding bugs are ones that already exist. However, I am going to
do a complete walkthrough later tomorrow, so hold off on going JIRA crazy if you see
anthing. I'm going to file JIRAs for:
- lack of icons for Search/Browse Workspaces links
- Moving workspaces does not work (this isn't a UI issue)
- buttons are weirdly formatted sometimes (filed this earlier)

The thing which I haven't sufficiently tested is the policy enforcement failures.