GALAXY-238: ArtifactView UI first cut.
Should be self evident to use. What doesn't work:
* Editing views
* Not all queries are correctly serialized yet.
* Styling of the search panel sucks. Need to bring it in line 
  with the main UI yet.

I refactored the Query/JcrRegistryImpl to make it easier
to constructor a query from a string and save it as part of this.

There is also some cleanup to AbstractBrowsePanel so that
the sub classes are responsible for fetching the artifacts.