GALAXY-238: Get Edit mode for views working.
Still some non functional UI CSS cleanup I would like to do.

Add some tests for parsing queries - != was not working correctly.

Move the appropriate Messages.properties messages into the api module
for the Query class.

Fix the artifact type select on the Search panel by creating
a getHistoryToken() method which returns the current token for that page
so the artifact types can link to it.