Introduce the concept of Entrys in the repository. These are exactly like an Artifact, except without the file/data.
This is useful if you just want to store metadata. For instance,
you can create an Entry which represents a Service. Or a machine.
Or an endpoint. Etc...

I was going to hold off on this until post 1.1, but since I 
was refactoring the Registry I figured I might as well at least
introduce this to the core now. This will save me from 
changing everything yet again down the line.